Refacctoring
